add unicode input creation
There where issues with some Windows applications like Notepad on a German keyboard layout. To type a `}` on the keyboard AltGr+0 is pressed. This is interpreted as Ctrl+Alt+0. But sending this with the SendInput function does not to work for Notepad.
The workaround is to use the unicode input directly where language dependent key mappings are not needed.